move home screen into own component
